+++ src/bika/lims/browser/pricelist.py
@@ -25,7 +25,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.api.security import check_permission
 from bika.lims.browser import BrowserView
 from bika.lims.browser.bika_listing import BikaListingView
-from bika.lims.permissions import AddPricelist
+from senaite.core.permissions import AddPricelist
 from bika.lims.utils import get_link
 from DateTime import DateTime
 from Products.Five.browser.pagetemplatefile import ViewPageTemplateFile