+++ src/bika/lims/controlpanel/bika_artemplates.py
@@ -27,7 +27,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import IARTemplates
-from bika.lims.permissions import AddARTemplate
+from senaite.core.permissions import AddARTemplate
 from bika.lims.utils import get_link
 from plone.app.content.browser.interfaces import IFolderContentsView
 from plone.app.folder.folder import ATFolder