+++ src/bika/lims/controlpanel/bika_labproducts.py
@@ -25,7 +25,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import ILabProducts
-from bika.lims.permissions import AddLabProduct
+from senaite.core.permissions import AddLabProduct
 from bika.lims.utils import get_link
 from plone.app.folder.folder import ATFolder
 from plone.app.folder.folder import ATFolderSchema