+++ src/bika/lims/controlpanel/bika_preservations.py
@@ -25,7 +25,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import IPreservations
-from bika.lims.permissions import AddPreservation
+from senaite.core.permissions import AddPreservation
 from bika.lims.utils import get_link
 from plone.app.folder.folder import ATFolder
 from plone.app.folder.folder import ATFolderSchema