+++ src/bika/lims/controlpanel/bika_referencedefinitions.py
@@ -25,7 +25,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import IReferenceDefinitions
-from bika.lims.permissions import AddReferenceDefinition
+from senaite.core.permissions import AddReferenceDefinition
 from bika.lims.utils import get_image
 from bika.lims.utils import get_link
 from bika.lims.utils import t