+++ src/bika/lims/controlpanel/bika_sampleconditions.py
@@ -27,7 +27,7 @@ from bika.lims import bikaMessageFactory as _
 from bika.lims.browser.bika_listing import BikaListingView
 from bika.lims.config import PROJECTNAME
 from bika.lims.interfaces import ISampleConditions
-from bika.lims.permissions import AddSampleCondition
+from senaite.core.permissions import AddSampleCondition
 from bika.lims.utils import get_link
 from plone.app.folder.folder import ATFolder
 from plone.app.folder.folder import ATFolderSchema